// frontend/website/src/types/api.ts
export interface RelatedBill {
  congress: number;
  type: string;
  number: number;
  title: string;
  relationshipType: string;
  latestAction?: {
    actionDate: string;
    text: string;
  };
}